+++ src/forr.rs
@@ -1,63 +1,45 @@
-use std::fmt::Display;
 use std::iter;
-use std::ops::{BitAndAssign, Range};
+use std::ops::BitAndAssign;
 
-use manyhow::{bail, error_message, Error, ErrorMessage, JoinToTokensError, Result, SpanRanged};
+use manyhow::{bail, error_message, ErrorMessage, JoinToTokensError, Result};
 use proc_macro2::{token_stream, Group, Ident, Literal, Span, TokenStream, TokenTree};
 use proc_macro_utils::{
     Delimited, TokenParser, TokenStream2Ext, TokenTree2Ext, TokenTreeLiteral, TokenTreePunct,
 };
-use quote::{quote, ToTokens};
+use quote::{format_ident, quote, ToTokens};
 
-use crate::iff::step_colon;
 use crate::Parser;
 
-pub fn step_star(input: &mut Parser) -> Result<()> {
-    if input
-        .next_if_each((TokenTree::is_dollar, TokenTree::is_asterix))
-        .or_else(|| input.next_if_each((TokenTree::is_pound, TokenTree::is_asterix)))
-        .is_none()
-    {
-        bail_optional_span!(input, "expected `$*` or `#*`");
-    } else {
-        Ok(())
-    }
-}
+mod input;
+use input::*;
 
-pub fn forr(input: TokenStream) -> Result {
-    let mut context = Context::default();
-    let mut input = input.parser();
-    // parse variables
-    let vars = Vars::parse(&mut input)?;
-    for var in vars.idents() {
-        context.push_key(var)?;
-    }
-    // ... in
-    unwrap_or!(_, input.next_keyword("in"), input, "`in`");
-    // ... in [...]
-    unwrap_or!(values, input.next_bracketed(), input, "`[...]`");
-    let mut values = values.stream().parser();
+mod parse;
+
+pub fn forr(
+    MacroInput {
+        vars,
+        values,
+        prefix,
+        loop_type,
+        body,
+    }: MacroInput,
+) -> Result {
+    let mut context = Context::new(vars.idents(), prefix)?;
+
+    let values = match values {
+        Values::List(values) => values,
+        Values::Function(function) => function.expand(),
+    };
+
+    let mut values = values.parser();
     while !values.is_empty() {
         context.new_row();
         vars.parse_values(&mut values, &mut context, None)?;
     }
 
-    if input.peek_tt_pound().is_some() {
-        context.prefix = TokenTree::is_pound;
-    }
-
-    // $*
-    if step_star(&mut input).is_ok() {
-        Ok(replace_loop(input, &context))
-    } else if step_colon(&mut input).is_ok() {
-        replace_outer(input, &context)
-    } else if let Some(token) = input.next() {
-        bail!(token, "expected `$*`, `$:`, `#*`, or `#:`")
-    } else {
-        bail!(
-            "unexpected end of macro invocation: expected to be followed by `$*`, `$:`, `#*`, or \
-             `#:`"
-        )
+    match loop_type {
+        LoopType::Star => Ok(replace_loop(body, &context)),
+        LoopType::Colon => replace_outer(body, &context),
     }
 }
 
@@ -66,46 +48,34 @@ struct Context {
     keys: Vec<Ident>,
     data: Vec<Value>,
     len: usize,
-    prefix: fn(&TokenTree) -> bool,
+    prefix: Prefix,
 }
 
-impl Default for Context {
-    fn default() -> Self {
-        Self {
-            keys: Vec::new(),
+impl Context {
+    fn new(vars: impl IntoIterator<Item = Ident>, prefix: Prefix) -> Result<Self> {
+        let mut keys = Vec::new();
+        for key in vars {
+            if keys.contains(&key) {
+                bail!(key, "variable `{key}` already used twice");
+            }
+            keys.push(key);
+        }
+        Ok(Self {
+            keys,
             data: Vec::new(),
             len: 0,
-            prefix: TokenTree::is_dollar,
-        }
+            prefix,
+        })
     }
 }
 
 impl Context {
-    fn push_key(&mut self, key: Ident) -> Result<()> {
-        assert_eq!(self.len, 0, "can only add keys before pushing values");
-
-        if let Some(present) = self.keys.iter().find(|k| *k == &key) {
-            Err(
-                error_message!(key, "variable `{key}` specified twice").join(error_message!(
-                    present,
-                    "variable `{present}` specified twice"
-                )),
-            )
-        } else {
-            self.keys.push(key);
-            Ok(())
-        }
-    }
-
     fn new_row(&mut self) {
         self.len += 1;
         self.data
             .reserve(self.len * self.keys.len() - self.data.len());
     }
 
-    // fn finish_row(&mut self) {
-    //     self.data.extend((self.data.len()..self.len * self.keys.len()).map(|_|
-    // Value::Optional(None))); }
     fn push(&mut self, value: Value) {
         self.data.push(value);
     }
@@ -129,7 +99,7 @@ fn replace_outer(input: impl Into<TokenParser>, context: &Context) -> Result {
             )));
             continue;
         }
-        let Some(prefix) = input.next_if(context.prefix) else {
+        let Some(prefix) = input.next_if(context.prefix.test()) else {
             output.extend(input.next());
             continue;
         };
@@ -187,7 +157,7 @@ fn replace_inner(
     let mut output = TokenStream::new();
     let mut tokens = tokens.into();
     while !tokens.is_empty() {
-        if tokens.peek_if_each(context.prefix).is_some() {
+        if tokens.peek_if_each(context.prefix.test()).is_some() {
             if let Some(ident) = tokens.peek_n_ident(1) {
                 if let Some(replacement) = context.get(index, ident) {
                     // $
@@ -239,16 +209,6 @@ fn replace_inner(
     (output, tl_some)
 }
 
-#[derive(Clone, Copy, PartialEq)]
-enum Typ {
-    Expr,
-    Ty,
-    Tt,
-    Inner,
-    Idx,
-    Ident,
-}
-
 impl Typ {
     fn parse_value(
         self,
@@ -278,6 +238,23 @@ impl Typ {
                 .next_ident()
                 .map(ToTokens::into_token_stream)
                 .ok_or_else(|| error_message!(input.peek(), "expected ident")),
+            Typ::Block => input
+                .next_braced()
+                .map(ToTokens::into_token_stream)
+                .ok_or_else(|| error_message!(input.peek(), "expected block")),
+            Typ::Path => input
+                .next_type()
+                .ok_or_else(|| error_message!(input.peek(), "expected path")),
+            Typ::Lifetime => input
+                .next_if_each((
+                    |tt: &TokenTree| tt.is_joint() && tt.is_quote(),
+                    TokenTree::is_ident,
+                ))
+                .ok_or_else(|| error_message!(input.peek(), "expected lifetime")),
+            Typ::Literal => input
+                .next_literal()
+                .map(ToTokens::into_token_stream)
+                .ok_or_else(|| error_message!(input.peek(), "expected literal")),
         }
     }
 
@@ -286,137 +263,13 @@ impl Typ {
     }
 }
 
-impl TryFrom<Ident> for Typ {
-    type Error = Error;
-
-    fn try_from(value: Ident) -> Result<Self> {
-        const ERROR: &str = "expected `expr`, `ty`, `tt` or `inner`";
-        const NOTE: &str = "this crates patterns are fuzzy, i.e. `ty` and `expr` can parse any \
-                            pattern matching the comma rules for type and expression";
-        let help = match value.to_string().as_str() {
-            "expr" => return Ok(Self::Expr),
-            "ty" => return Ok(Self::Ty),
-            "tt" => return Ok(Self::Tt),
-            "inner" => return Ok(Self::Inner),
-            "idx" => return Ok(Self::Idx),
-            "block" => "`tt`",
-            "ident" => return Ok(Self::Ident),
-            "path" => "`ty`",
-            "lifetime" | "literal" => "`ty` or `expr`",
-            "vis" => "`ty?` or `expr?`",
-            _ => bail!(value, "{ERROR}";
-                note = "{NOTE}";
-                help = "use `inner` instead and wrap the value in `(...)`"
-            ),
-        };
-        bail!( value, "{ERROR}";
-            note = "{NOTE}";
-            help = "use {help} instead"
-        );
-    }
-}
-
-enum Vars {
-    Single(Var),
-    Plain(Vec<Vars>),
-    Tuple(Vec<Vars>, Span),
-    Function {
-        function: Function,
-        span: Range<Span>,
-        optional: bool,
-    },
-}
-
-impl SpanRanged for Vars {
-    fn span_range(&self) -> Range<proc_macro2::Span> {
-        match self {
-            Vars::Plain(vars) => {
-                vars.first()
-                    .expect("vars are nether empty")
-                    .span_range()
-                    .start
-                    ..vars.last().expect("vars are nether empty").span_range().end
-            }
-            Vars::Tuple(_, span) => span.span_range(),
-            Vars::Single(var) => var.span_range(),
-            Vars::Function { span, .. } => span.span_range(),
-        }
-    }
-}
-
 impl Vars {
-    fn optional(&self) -> bool {
+    fn non_consuming(&self) -> bool {
         match self {
-            Vars::Single(var) => var.optional,
-            Vars::Plain(vars) => vars.iter().all(Vars::optional),
-            Vars::Tuple(..) => false,
-            Vars::Function { optional, .. } => *optional,
+            Vars::Single(v) => matches!(v.typ, Typ::Idx),
+            Vars::Plain(v) => v.iter().all(Self::non_consuming),
+            Vars::Tuple(..) | Vars::Function { .. } => false,
         }
-    }
-
-    fn idx(&self) -> bool {
-        matches!(self, Self::Single(Var { typ: Typ::Idx, .. }))
-    }
-
-    fn parse_multiple(input: &mut Parser, top_level: bool) -> Result<Vec<Self>> {
-        let mut vars = Vec::new();
-        while !(input.is_empty() || (top_level && input.peek_keyword("in").is_some())) {
-            // $ or #
-            if input.next_tt_dollar().is_some() || input.next_tt_pound().is_some() {
-                vars.push(Var::parse(input).map(Self::Single)?);
-            } else if let Some(input) = input.next_parenthesized() {
-                vars.push(Vars::Tuple(
-                    Vars::parse_multiple(&mut input.stream().parser(), false)?,
-                    input.span(),
-                ));
-            } else {
-                let span = input.span();
-                unwrap_or!(
-                    name,
-                    input.next_ident(),
-                    input,
-                    "either `$ident`, `(...)` or `casing(...)`"
-                );
-                let mut span = span..input.span();
-                unwrap_or!(args, input.next_parenthesized(), input, "`(...)`");
-
-                let optional = input.next_if(TokenTree::is_question).map(|t| t.span());
-                span.end = optional.unwrap_or(span.end);
-
-                vars.push(Vars::Function {
-                    function: Function::parse(&name, args.stream().parser(), optional)?,
-                    optional: optional.is_some(),
-                    span,
-                });
-            }
-            if input.is_empty() || (top_level && input.peek_keyword("in").is_some()) {
-                break;
-            }
-            if input.next_tt_comma().is_none() {
-                bail!(input.into_token_stream(), "expected `,`");
-            }
-        }
-
-        Ok(vars)
-    }
-
-    fn parse(input: &mut Parser) -> Result<Self> {
-        let vars = Vars::parse_multiple(input, true)?;
-        if vars.is_empty() {
-            bail_optional_span!(input, "expected binding, either $ident:type or (...)");
-        }
-        let vars = Self::Plain(vars);
-        if vars.optional() {
-            bail!(vars, "top level binding cannot be optional");
-        }
-        if vars.idx() {
-            bail!(
-                vars,
-                "at least one consuming binding must be specified i.e. of type `expr`, `ty`, `tt` \
-                 or `inner`"
-            );
-        }
-        Ok(vars)
     }
 
     fn parse_values(
@@ -431,13 +284,11 @@ impl Vars {
                     vars.parse_values(input, context, group_span)?;
                 }
             }
-            Vars::Tuple(vars, _) => {
+            Vars::Tuple(vars) => {
                 let group_span = input.peek().map(TokenTree::span);
                 unwrap_or!(inner, input.next_parenthesized(), input, "(...)");
                 let mut inner = inner.stream().parser();
-                for vars in vars {
-                    vars.parse_values(&mut inner, context, group_span)?;
-                }
+                vars.parse_values(&mut inner, context, group_span)?;
 
                 if !input.is_empty() {
                     unwrap_or!(_, input.next_tt_comma(), input, "`,`");
@@ -468,8 +319,8 @@ impl Vars {
                 }
             }
             Vars::Function {
+                name,
                 function,
-                span,
                 optional,
             } => {
                 if input.is_empty() {
@@ -479,8 +330,8 @@ impl Vars {
                         }
                     } else {
                         bail!(
-                            error_message!(group_span, "missing mandatory values for `{function}`")
-                                .join(error_message!(span, "`{function}` defined mandatory here"))
+                            error_message!(group_span, "missing mandatory values for `{name}`")
+                                .join(error_message!(name, "`{name}` defined mandatory here"))
                         );
                     }
                 } else {
@@ -496,69 +347,13 @@ impl Vars {
         Ok(())
     }
 
-    fn idents(&self) -> impl Iterator<Item = Ident> + '_ {
-        let box_: Box<dyn Iterator<Item = Ident>> = match self {
+    fn idents(&self) -> Box<dyn Iterator<Item = Ident> + '_> {
+        match self {
             Vars::Single(Var { name, .. }) => Box::new(iter::once(name.clone())),
-            Vars::Tuple(vars, _) | Vars::Plain(vars) => {
-                Box::new(vars.iter().flat_map(Vars::idents))
-            }
+            Vars::Tuple(vars) => Box::new(vars.idents()),
+            Vars::Plain(vars) => Box::new(vars.iter().flat_map(Vars::idents)),
             Vars::Function { function, .. } => Box::new(function.idents()),
-        };
-        box_
-    }
-}
-
-#[derive(Clone)]
-struct Var {
-    name: Ident,
-    typ: Typ,
-    optional: bool,
-    span: Range<Span>,
-}
-
-impl SpanRanged for Var {
-    fn span_range(&self) -> Range<Span> {
-        self.span.clone()
-    }
-}
-
-impl Var {
-    fn parse(input: &mut Parser) -> Result<Self> {
-        let span = input.span();
-        // $ident
-        unwrap_or!(name, input.next_ident(), input, "ident");
-        // $ident:
-        unwrap_or!(_, input.next_tt_colon(), input, "`:`");
-        let mut span = span..input.span();
-        // $ident:typ
-        unwrap_or!(
-            typ,
-            input.next_ident(),
-            input,
-            "`expr`, `ty`, `tt` or `inner`"
-        );
-        let q_span = input.span();
-        let optional = if let Some(question) = input.next_tt_question() {
-            span.end = q_span;
-            if typ == "tt" {
-                bail!(
-                    quote!(#typ #question),
-                    "`tt` cannot be optional due to parsing instructions use `inner` instead and \
-                     wrap the value in `(...)`"
-                )
-            }
-            true
-        } else {
-            false
-        };
-        let typ = Typ::try_from(typ)?;
-
-        Ok(Self {
-            name,
-            typ,
-            optional,
-            span,
-        })
+        }
     }
 }
 
@@ -585,35 +380,16 @@ impl Value {
     }
 }
 
-enum Function {
-    Casing(Vec<(Ident, Casing)>),
-}
-
-impl Display for Function {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Function::Casing(v) => write!(
-                f,
-                "casing({})",
-                v.iter()
-                    .map(|(i, c)| format!("{i}:{c:?}"))
-                    .collect::<Vec<_>>()
-                    .join(",")
-            ),
-        }
-    }
-}
-
-impl Function {
+impl VarsFunction {
     fn len(&self) -> usize {
         match self {
-            Function::Casing(v) => v.len(),
+            Self::Casing(v) => v.len(),
         }
     }
 
     fn parse_value(&self, input: &mut TokenParser) -> Result<Vec<TokenStream>> {
         match self {
-            Function::Casing(variables) => {
+            Self::Casing(variables) => {
                 let mut span = None;
                 let parts: Vec<_> = input
                     .next_while(|token| token.is_ident() || token.is_literal())
@@ -664,82 +440,51 @@ impl Function {
 
     fn idents(&self) -> impl Iterator<Item = Ident> + '_ {
         match self {
-            Function::Casing(c) => c.iter().map(|i| i.0.clone()),
-        }
-    }
-
-    fn parse(name: &Ident, mut args: TokenParser, _optional: Option<Span>) -> Result<Self> {
-        match name.to_string().as_str() {
-            "casing" => {
-                let mut casings = Vec::new();
-                while !args.is_empty() {
-                    unwrap_or!(
-                        _,
-                        args.next_tt_dollar().or_else(|| args.next_tt_pound()),
-                        args,
-                        "`$`"
-                    );
-                    unwrap_or!(name, args.next_ident(), args, "ident");
-                    unwrap_or!(_, args.next_tt_colon(), args, ":");
-                    unwrap_or!(
-                        casing,
-                        args.next_ident(),
-                        args,
-                        "either `c`, `C`, `s` or `S`"
-                    );
-                    let casing = match casing.to_string().as_str() {
-                        "c" => Casing::c,
-                        "C" => Casing::C,
-                        "s" => Casing::s,
-                        "S" => Casing::S,
-                        other => bail!(
-                            name,
-                            "unknown case `{other}`, expected either `c`, `C`, `s` or `S`"
-                        ),
-                    };
-                    casings.push((name, casing));
-                    if !args.is_empty() {
-                        unwrap_or!(_, args.next_tt_comma(), args, "`,`");
-                    }
-                }
-                Ok(Self::Casing(casings))
-            }
-            other => bail!(name, "unknown function `{other}`, expected `casing`"),
+            Self::Casing(c) => c.iter().map(|i| i.0.clone()),
         }
     }
 }
 
-#[allow(non_camel_case_types)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum Casing {
-    s,
-    S,
-    c,
-    C,
+impl ValuesFunction {
+    fn expand(self) -> TokenStream {
+        match self {
+            ValuesFunction::Ident(prefix, count) => (0..count)
+                .map(|idx| {
+                    let ident = format_ident!("{prefix}{idx}");
+                    quote!(#ident,)
+                })
+                .collect(),
+        }
+    }
+}
+
+#[cfg(test)]
+macro_rules! assert_expansion {
+    ($in:tt, $out:tt) => {
+        proc_macro_utils::assert_tokens!(forr(syn::parse_quote!$in).unwrap(), $out);
+    };
 }
 
 #[test]
 fn expansions() {
-    use proc_macro_utils::assert_expansion;
-    assert_expansion!(forr! {$a:expr in [1, 2, 3]$* $a}.unwrap(), {1 2 3});
-    assert_expansion!(forr! {$a:expr, $i:idx in [1, 2, 3]$* $i}.unwrap(), {0 1 2});
-    assert_expansion!(forr! {$a:expr, $i:idx in [1, 2, 3]$: ! $($i)* !}.unwrap(), {! 0 1 2 !});
-    assert_expansion!(forr! {$a:expr?, $i:idx in [1, 2, 3]$: ! $($i)* !}.unwrap(), {! 0 1 2 !});
-    assert_expansion!(forr! {($a:expr?), $i:idx in [(1), (), (3)]$: ! $($i$(:$a)?)* !}.unwrap(), {! 0:1 1 2:3 !});
+    assert_expansion!({$a:expr in [1, 2, 3]$* $a}, {1 2 3});
+    assert_expansion!({$a:expr, $i:idx in [1, 2, 3]$* $i}, {0 1 2});
+    assert_expansion!({$a:expr, $i:idx in [1, 2, 3]$: ! $($i)* !}, {! 0 1 2 !});
+    assert_expansion!({$a:expr, $i:idx in [1, 2, 3]$: ! $($i)* !}, {! 0 1 2 !});
+    assert_expansion!({($a:expr?), $i:idx in [(1), (), (3)]$: ! $($i$(:$a)?)* !}, {! 0:1 1 2:3 !});
     // This behaviour might not be the expected one, but it enables better support
     // for nested forr loops, as the groups of an inner forr loop are left
     // untouched.
     assert_expansion!(
-        forr! {$a:expr, $b:expr?, $i:idx in [1a, 1b, 2a]$: ! $($i:$($a)? $()? $(+ $b)? $()* )* !}.unwrap(),
+        {($a:expr, $b:expr?), $i:idx in [(1a, 1b), (2a)]$: ! $($i:$($a)? $()? $(+ $b)? $()* )* !},
         {! 0:$(1a)? $()? + 1b $()* 1:$(2a)? $()? $()* !}
     );
 }
 
 #[test]
 fn casing() {
-    use proc_macro_utils::assert_expansion;
     assert_expansion!(
-        forr! {casing($s:s, $S:S, $c:c, $C:C) in [a b, "a b", a 1, a 0x23]$* $s $S $c $C }.unwrap(),
+        {casing($s:s, $S:S, $c:c, $C:C) in [a b, "a b", a 1, a 0x23]$* $s $S $c $C },
         {
             a_b    A_B    aB    AB
             a_b    A_B    aB    AB
@@ -747,6 +492,11 @@ fn casing() {
             a_0x23 A_0X23 a0x23 A0x23
         }
     );
-    assert_expansion!(forr! {casing($s:s,) in [a b] $* $s}.unwrap(), { a_b });
-    assert_expansion!(forr! {($a:expr, casing($s:s)?) in [(a, b c), (d)] $* $a $s}.unwrap(), {a b_c d});
+    assert_expansion!({casing($s:s,) in [a b] $* $s}, { a_b });
+    assert_expansion!({($a:expr, casing($s:s)?) in [(a, b c), (d)] $* $a $s}, {a b_c d});
+}
+
+#[test]
+fn idents() {
+    assert_expansion!({$a:ident in idents(a, 2) $* $a}, {a0 a1});
 }